Added pre-Filtering to TNTClusterer and ScanClusterer
