+++ TrkHitReco/fcl/prolog.fcl
@@ -41,11 +41,27 @@ TNTClusterer : {
     TimeRMS          : 2.0
     DeltaTimeBinMin  : 10
     MedianCentroid   : false
-    preFilter        : true
-    pfTimeBin        : 20 
+    preFilter        : false
+    pfTimeBin        : 20
     pfPhiBin         : 0.1
-    pfMinHit         : 5
+    pfMinHit         : 10
+    pfMinSumHit      : 20
     ComboInit        : true
+    TestFlag         : true
+    BackgroundMask   : []
+    SignalMask       : ["TimeSelection", "EnergySelection","RadiusSelection"]
+}
+
+ScanClusterer: {
+    Pbin             : 0.1
+    Tbin             : 20
+    Rbin             : 50 
+    Rmin             : 350
+    Rmax             : 650
+    MinPeakHit       : 2
+    MinSeedHit       : 3
+    MinRadHit        : 50
+    FilterAlgo       : 1
     TestFlag         : true
     BackgroundMask   : []
     SignalMask       : ["TimeSelection", "EnergySelection","RadiusSelection"]
@@ -58,6 +74,7 @@ FlagBkgHits : {
   StrawHitCollection   : "makeSH"
   ClusterAlgorithm     : 1
   TNTClustering        : { @table::TNTClusterer }
+  ScanClustering       : { @table::ScanClusterer }
   MinActiveHits        : 3
   MinNPlanes           : 2
   ClusterPositionError : 10
